Nil guards in flow-sensitive typing
